Fix a bug where storage was infinitely being restored (#369)
* make tasks wait until storage fully restored

* queue should be depleted before restore

* revert log for message count

* rustfmt

* node-template compile

* add comment explaining unused

* propogate errors from storage restore

+++ work-queue/sa-work-queue/src/runner.rs
@@ -223,8 +223,8 @@ impl<Env: 'static> Runner<Env> {
 		self.threadpool.queued_count()
 	}
 
-	pub fn job_count(&self) -> usize {
-		self.handle.queue.message_count() as usize
+	pub fn job_count(&self) -> u32 {
+		self.handle.queue.message_count()
 	}
 
 	pub fn max_jobs(&self) -> usize {
